+++ packages/internal-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @defuse-protocol/internal-utils
 
+## 0.37.0
+
+### Minor Changes
+
+- 2073fd5: Add ADI support to PoA Bridge while retaining native ADI support through HOT Bridge.
+
+## 0.36.0
+
+### Minor Changes
+
+- 8975fe8: Block withdrawals to token addresses
+
 ## 0.35.2
 
 ### Patch Changes